Move system_instruction tests into test_get_llm_invocation_params.py
